django/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --format requirements-txt
asgiref==3.11.1
    # via django
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
    # via cryptography
cryptography==46.0.5
    # via anon-confession-backend
django==6.0.3
    # via
    #   anon-confession-backend
    #   djangorestframework
django-ipware==7.0.1
    # via anon-confession-backend
django-ratelimit==4.1.0
    # via anon-confession-backend
djangorestframework==3.16.1
    # via anon-confession-backend
pycparser==3.0 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
    # via cffi
pymysql==1.1.2
    # via anon-confession-backend
python-dotenv==1.2.2
    # via anon-confession-backend
python-ipware==3.0.0
    # via django-ipware
sqlparse==0.5.5
    # via django
tzdata==2025.3 ; sys_platform == 'win32'
    # via django
